Remove async image fetch methods, update context references
The async functions responsible for fetching different types of images have been removed. Additionally, in various .rs files within the src/commands/slash_commands directory, the `ctx` variable's reference is updated in functions from `&Context` to `Context`. The primary modification is the change in reference for the `ctx` variable to support correct interaction with the 'serenity' library's Context structure in asynchronous runtime conditions.